Delegate Req.Steps.auth,handle_http_digest to Req.Auth
